+++ src/main/java/com/dchprojects/mydictionaryrestapi/repository/LanguageRepository.java
@@ -6,6 +6,6 @@ import com.dchprojects.mydictionaryrestapi.entity.LanguageEntity;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LanguageRepository extends JpaRepository<LanguageEntity, Integer> {
+public interface LanguageRepository extends JpaRepository<LanguageEntity, Long> {
 
 }
